# openapi.json
{
  "components": {
    "schemas": {
...
      "ChapterArchiveRequest": {
        "description": "Optional reassignment targets when archiving a chapter.\nBoth default to None \u2014 rows that aren't reassigned simply\nstay linked to the soft-deleted chapter (events become\ninvisible until restore, users still have a chapter_id\npointing at the archived chapter).",
        "properties": {
          "reassign_events_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reassign Events To"
          },
          "reassign_users_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reassign Users To"
          }
        },
        "title": "ChapterArchiveRequest",
        "type": "object"
      },
      "ChapterCreate": {
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ChapterCreate",
        "type": "object"
      },
      "ChapterOut": {
        "description": "One chapter, live or archived. ``archived`` mirrors\n``deleted_at IS NOT NULL`` so the frontend doesn't need to\ninspect timestamps.",
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "city_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Lat"
          },
          "city_lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Lon"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "archived",
          "city",
          "city_lat",
          "city_lon"
        ],
        "title": "ChapterOut",
        "type": "object"
      },
      "ChapterPatch": {
        "description": "Partial update for a chapter. Either field may be omitted;\nthe city tuple is all-or-nothing \u2014 passing ``city`` requires\n``city_lat``/``city_lon`` (and vice versa) so we never end up\nwith a city display name that has no coordinates.",
        "properties": {
          "city": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "city_lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Lat"
          },
          "city_lon": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Lon"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ChapterPatch",
        "type": "object"
      },
...
      "ChapterUsageOut": {
        "description": "How many users / events are currently assigned to this\nchapter. Used by the delete dialog so the admin sees what\nhappens to dependents before confirming.",
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "users": {
            "title": "Users",
            "type": "integer"
          }
        },
        "required": [
          "users",
          "events"
        ],
        "title": "ChapterUsageOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Opkomst",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chapters": {
      "get": {
        "description": "List chapters. By default only active ones; pass\n``include_archived=true`` to also surface soft-deleted ones (used\nby the admin autocomplete to support restore).",
        "operationId": "list_chapters_api_v1_chapters_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterOut"
                  },
                  "title": "Response List Chapters Api V1 Chapters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chapters",
        "tags": [
          "chapters"
        ]
      },
      "post": {
        "operationId": "create_chapter_api_v1_chapters_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
    "/api/v1/chapters/{chapter_id}": {
      "delete": {
        "operationId": "archive_chapter_api_v1_chapters__chapter_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ChapterArchiveRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Chapter",
        "tags": [
          "chapters"
        ]
      },
      "patch": {
        "description": "Partial update \u2014 name and/or city. The city tuple is set\ntogether (display name + lat + lon); leaving all three at\n``None`` in the payload clears a previously-set city. Pass\nnothing for ``city*`` keys to leave the current city\nuntouched.",
        "operationId": "patch_chapter_api_v1_chapters__chapter_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
    "/api/v1/chapters/{chapter_id}/restore": {
      "post": {
        "operationId": "restore_chapter_api_v1_chapters__chapter_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
    "/api/v1/chapters/{chapter_id}/usage": {
      "get": {
        "description": "How many live users + events are currently linked to this\nchapter. The frontend calls this before opening the delete\ndialog so it can offer reassignment dropdowns when there's\nsomething at stake.",
        "operationId": "chapter_usage_api_v1_chapters__chapter_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterUsageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chapter Usage",
        "tags": [
          "chapters"
        ]
      }
    },
...
  }
}